+++ web/modules/custom/hoeringsportal_activity/assets/js/course-form.js
@@ -1,0 +1,27 @@
+let fromTimeField = document.querySelector('#edit-field-first-meeting-time-wrapper .form-time');
+let toTimeField = document.querySelector('#edit-field-last-meeting-time-wrapper .form-time');
+let hideTimeCheckbox = document.querySelector('#edit-field-hide-time-value');
+
+hideTimeCheckbox.addEventListener('click', function(event) {
+  fromTimeField.classList.toggle('hidden')
+  toTimeField.classList.toggle('hidden')
+  if (fromTimeField.classList.contains('hidden')) {
+    fromTimeField.value = '00:00';
+    toTimeField.value = '23:59';
+  }
+  else {
+    const now = new Date();
+    const currentHours = String(now.getHours()).padStart(2, '0');
+    const currentMinutes = String(now.getMinutes()).padStart(2, '0');
+    const formattedTime = currentHours + ':' + currentMinutes;
+    fromTimeField.value = formattedTime;
+    toTimeField.value = formattedTime;
+  }
+})
+
+if (hideTimeCheckbox.checked) {
+  fromTimeField.classList.toggle('hidden')
+  toTimeField.classList.toggle('hidden')
+  fromTimeField.value = '00:00';
+  toTimeField.value = '23:59';
+}