+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-586](https://github.com/itk-dev/hoeringsportal/pull/586)
+  * Added course content type
+  * Added activity wrapper for public_events and courses
 * [PR-576](https://github.com/itk-dev/deltag.aarhus.dk/pull/576)
   Updates translations and added translations handling to deployment
 * [PR-584](https://github.com/itk-dev/deltag.aarhus.dk/pull/584)